refactoring del codice, rimosse credenziali al login
