+++ frontend/src/components/LoginDialog.vue
