+++ frontend/src/components/layout/AppFooter.vue
@@ -1,5 +1,5 @@
 <script setup>
-import { useLayout } from '@/layout/composables/layout';
+import { useLayout } from '@/components/layout/composables/layout';
 import { computed } from 'vue';
 
 const { layoutConfig } = useLayout();
