+++ frontend/src/components/layout/AppTopbar.vue
@@ -1,6 +1,6 @@
 <script setup>
 import { ref, computed, onMounted, onBeforeUnmount } from 'vue';
-import { useLayout } from '@/layout/composables/layout';
+import { useLayout } from '@/components/layout/composables/layout';
 import { useRouter } from 'vue-router';
 import AuthService from '@/services/auth.service';
