--- frontend/src/layout/AppSidebar.vue
@@ -1,9 +1,0 @@
-<script setup>
-import AppMenu from './AppMenu.vue';
-</script>
-
-<template>
-    <app-menu></app-menu>
-</template>
-
-<style lang="scss" scoped></style>
