--- frontend/src/views/Log.vue
@@ -1,10 +1,0 @@
-<script setup>
-import { onMounted, reactive, ref, watch } from 'vue';
-import { useLayout } from '@/layout/composables/layout';
-
-const { isDarkTheme } = useLayout();
-
-</script>
-<template>
-    <h1>Log</h1>
-</template>
